TackVirtualMachine uses a 32-bit word size now
This means that registers are natively 32-bits wide and pointers are
natively 32-bits wide. However, we do not have Tack instructions for
manipulating 32-bit values so this is of limited utility right now.

+++ TurtleTools/SnapCore/TackVirtualMachine.swift
@@ -20,7 +20,7 @@ public enum TackVirtualMachineError: Error {
 
 public class TackVirtualMachine: NSObject {
     public typealias Register = TackInstruction.Register
-    public typealias Word = UInt16
+    public typealias Word = UInt32
     
     public let kMemoryMappedSerialOutputPort: Word = 0x0001
     public let kPageSize = 4096
